+++ go/src/koding/kites/kloud/provider/koding/machine.go
@@ -4,6 +4,7 @@ import (
 	"koding/db/models"
 	"koding/kites/kloud/contexthelper/session"
 	"koding/kites/kloud/eventer"
+	"koding/kites/kloud/klient"
 	"koding/kites/kloud/machinestate"
 	"koding/kites/kloud/plans"
 	"time"
@@ -138,4 +139,22 @@ func (m *Machine) markAsNotInitialized() error {
 			}},
 		)
 	})
+}
+
+func (m *Machine) isKlientReady() bool {
+	m.Log.Debug("All finished, testing for klient connection IP [%s]", m.IpAddress)
+	klientRef, err := klient.NewWithTimeout(m.Session.Kite, m.QueryString, time.Minute*5)
+	if err != nil {
+		m.Log.Warning("Connecting to remote Klient instance err: %s", err)
+		return false
+	}
+	defer klientRef.Close()
+
+	m.Log.Debug("Sending a ping message")
+	if err := klientRef.Ping(); err != nil {
+		m.Log.Debug("Sending a ping message err: %s", err)
+		return false
+	}
+
+	return true
 }
